# refactor(container,lazy,hooks): avoid blocking JS thread by refactoring SharedValue.value access