--- a/example/src/Shared/ExampleMasonry.tsx
+++ b/example/src/Shared/ExampleMasonry.tsx
@@ -69,11 +69,7 @@ const ItemSeparator = () => <View style={styles.separator} />
 const ListEmptyComponent = () => {
   const { top, height } = Tabs.useHeaderMeasurements()
   const translateY = useDerivedValue(() => {
-    return interpolate(
-      -top.value,
-      [0, height.value || 0],
-      [-(height.value || 0) / 2, 0]
-    )
+    return interpolate(-top.value, [0, height || 0], [-(height || 0) / 2, 0])
   }, [height])
 
   const stylez = useAnimatedStyle(() => {