--- a/src/ScrollView.tsx
+++ b/src/ScrollView.tsx
@@ -6,7 +6,6 @@ import {
   useAfterMountEffect,
   useChainCallback,
   useCollapsibleStyle,
-  useConvertAnimatedToValue,
   useScrollHandlerY,
   useSharedAnimatedRef,
   useTabNameContext,
@@ -91,16 +90,14 @@ export const ScrollView = React.forwardRef<
       [progressViewOffset, refreshControl]
     )
 
-    const contentInsetValue = useConvertAnimatedToValue(contentInset)
-
     const memoContentInset = React.useMemo(
-      () => ({ top: contentInsetValue }),
-      [contentInsetValue]
+      () => ({ top: contentInset }),
+      [contentInset]
     )
 
     const memoContentOffset = React.useMemo(
-      () => ({ x: 0, y: -contentInsetValue }),
-      [contentInsetValue]
+      () => ({ x: 0, y: -contentInset }),
+      [contentInset]
     )
 
     const memoContentContainerStyle = React.useMemo(