Hey  @andreialecu  and @gkartalis (tagging bc I can't add reviewers), I was introducing this library to solve a UI pattern and noticed blocking JavaScript thread behavior that caused a regression of over 500ms to our initial screen render. 

<img width="800" alt="cpu-profile-of-collapsible-tab-view" src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/548a5b44-f037-44ce-ad27-57c376d2bd88">

The flame graph above shows that the JavaScript thread is blocked by executeOnUIRuntimeSync for quite some time impacting the initial render time of a screen by over 500ms.

On closer analysis I found that the problem is caused by reading a SharedValue.value in the main JS thread. This library does this quite often in places that could avoid it. This PR refactors most of the places where it does not need to be a SharedValue while trying to not cause any regressions.

I understand there isn't unit or e2e tests for this library so I tried my best to test the Example app. I posted multiple recordings below of how it currently works with these changes.

[I also wrote a detailed blogpost explaining this Reanimated blocking behavior](https://andrei-calazans.com/posts/reanimated-blocking-js-thread/).

You can also inspect the before and after CPU profilings, [I stored them here](https://gist.github.com/AndreiCalazans/3fc4a522f94ea3387060f9c6fe461840/revisions). The CPU profile will show that before these changes there were 57 calls to executeOnUIRuntimeSync when navigating from the main menu to the FlashList example in the Example app and now there are only 14 calls.

| Before (57 calls) | After (14 calls) |
|--|--|
| <img width="600" alt="image" src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/fbb2510f-21b9-402e-86fb-2b2c21892d46"> | <img width="600" alt="image" src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/d9f4b860-bd2a-4ea5-b3fd-407636e9e52a"> |


## Demos

| One | Two |
|--|--|
| <video src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/1f688b19-a427-428d-8fb8-0141d7f862ab"/> | <video src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/6b1b8e30-a7b9-4459-b83a-2175df02b4cc"/>  |

| Three | Four |
|--|--|
| <video src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/fd42dc3a-ef9b-4647-8d0e-3ab3eea26a24"/> | <video src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/40b58e5e-20a4-49e7-b357-0675da0d0cff"/>  |

| Five | Six |
|--|--|
| <video src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/ac4e036f-a7f0-47bc-b824-ce7411b69bb5"/> | <video src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/779df616-5b09-4d18-8f1a-79bcb6ae5eba"/>  |


| Seven | Eight |
|--|--|
| <video src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/130b7eaf-6245-4cdb-9451-e6e8f4256889"/> | <video src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/8615933a-4e99-4026-8b1f-5a9338150602"/>  |