@andreialecu 2024-06-19
Very interesting, I didn't know about this behavior of Reanimated. Is it documented anywhere?

Could it be a bug in Reanimated, maybe a recent regression? I don't remember running into an issue before or hearing about SharedValue blocking JS threads.

@piaskowyk could you kindly shed some light? See Andrei Calazan's blog post and reddit posts:
https://andrei-calazans.com/posts/reanimated-blocking-js-thread/
https://www.reddit.com/r/reactnative/comments/1diy3pc/reanimated_can_block_the_main_javascript_thread/

I remember introducing `SharedValue` here specifically to improve performance in the past. For example to be able to do this on the UI thread:
https://github.com/PedroBern/react-native-collapsible-tab-view/compare/v5.0.0-rc.1...v5.0.0-rc.10#diff-739a99e40f851900557fbc98026c870d8ac3c2da65ec6fc29d60c161effaf20dR240-R248