@AndreiCalazans 2024-06-19
Hey @andreialecu, thanks for reviewing. 

I also didn't know this until now, not sure if it's a regression since the behavior kinda of makes sense if you think about it. Any time you need to read a SharedValue.value from within the JavaScript thread it needs to bridge the value synchronously from the HostObject so that can have a non-trivial cost (anywhere from 5 to 20ms depending on device).

I want to highlight that the problem is not using SharedValue but repeatedly needing to read it from within the JS thread. The [example diff you shared](https://github.com/PedroBern/react-native-collapsible-tab-view/commit/9378af6a0f95ce59a3eab9ca786baa05bb8039ba#diff-739a99e40f851900557fbc98026c870d8ac3c2da65ec6fc29d60c161effaf20dR240-R248) is a good example of the correct pattern which is to take something from the JS thread and run it in worklet to save CPU cost in the JS thread.