@andreialecu 2024-06-19
Are you using the new architecture?

With Paper/old architecture, as far as I remember, calling `SharedValue.value` in the past did not guarantee you'd get the actual last value. It was sometimes out of sync with the value from a worklet. Hence why I needed to create the `useConvertAnimatedToValue` hook, to try to ensure the value is kept in sync as much as possible.