@AndreiCalazans 2024-06-19
@andreialecu I'm not using the new arch nor did I test in it.