@AndreiCalazans 2024-06-19
@andreialecu note that my changes avoids this interchange of SharedValue to regular React state and solely relies on React state where possible.