@AndreiCalazans 2024-06-19
@andreialecu I understand the pushback. It will be helpful to hear from @piaskowyk to see if my understanding is correct. 

I think the Example app is enough of a minimal repro. I [pushed here](https://github.com/PedroBern/react-native-collapsible-tab-view/pull/424) a simple setup that allows us to run performance tests using Flashlight. 

I ran [these e2e](https://github.com/PedroBern/react-native-collapsible-tab-view/pull/424/files) tests with 10 iterations for `main` and for this branch. All tests were ran on the same Samsung a23 device (considered low end). These are the results:

<img width="1200" alt="Screenshot 2024-06-19 at 13 02 53" src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/0dcaee1b-b524-4edc-ab55-6a0152bb9dc3">

At a first glance it seems like nothing has changed and this branch is just slightly better, but take a look at when the FPS starts dipping when we navigate to the FlashList example. On `main` the dip happens at around 3 seconds for the 10 iterations while on the refactored branch it happens around 2.5 seconds this means this branch is half a second faster to load the components and dispatch the paint to UI natively. 

| Before | After |
|--|--|
| <img width="800" alt="Screenshot 2024-06-19 at 13 10 08" src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/9139133d-8457-4c7b-96c1-1aafd685c57d"> | <img width="800" alt="Screenshot 2024-06-19 at 13 02 53" src="https://github.com/PedroBern/react-native-collapsible-tab-view/assets/20777666/a3ffb8e9-8ff6-42dd-9385-4ee3f8c421be"> |


Flashlight reports:
- [Main branch](https://github.com/user-attachments/files/15903609/com.rnctvexample.json)
- [andrei/optimize-tab-view branch](https://github.com/user-attachments/files/15903608/flashlight_report_with_changes.json)

To visualize these yourself install Flashlight and run `flashlight report fileone filetwo`