@andreialecu 2024-06-21
I released this as 8.0.1-rc.0 to npm for the time being, so that we can all test it more easily.